tests/xno_gemini/requirements.txt
browser_use[all]==0.4.0
python_dotenv==1.1.1
allure_pytest==2.14.3
pytest_asyncio==1.0.0